Render update card on add-on page
